+++ ui/verifiedContracts/VerifiedContractsListItem.tsx
@@ -3,9 +3,9 @@ import BigNumber from 'bignumber.js';
 import React from 'react';
 
 import type { VerifiedContract } from 'types/api/contracts';
-import { CONTRACT_LICENSES } from 'types/client/contract';
 
 import config from 'configs/app';
+import { CONTRACT_LICENSES } from 'lib/contracts/licenses';
 import dayjs from 'lib/date/dayjs';
 import { currencyUnits } from 'lib/units';
 import CopyToClipboard from 'ui/shared/CopyToClipboard';
